used reduce EMD, the EMD is too much expensive
