--- src/main/java/com/telmomenezes/jfastemd/Feature.java
@@ -1,5 +1,0 @@
-package com.telmomenezes.jfastemd;
-
-public interface Feature {
-    public double groundDist(Feature f);
-}
